requirements.txt
seeed-python-grove
seeed-python-dht
RPi.GPIO
influxdb
influxdb-client
python-dotenv
pyyaml
apscheduler
sqlite-utils
pytest
